requirements.txt
Flask
transformers
whisper
pytesseract
pydub
requests
bs4
opencv-python
numpy
Pillow
